Move the SecureVaultCryptoProvider protocol to SecureStorage.
